requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
twilio==8.10.3
anthropic==0.7.8
pydantic==2.5.0
APScheduler==3.10.4
pytz==2023.3

